Prepare the metrics scraper for the bridge monitoring
We refactor the structure of the metrics scrapper to add more monitoring routines beyond the nodes observation job.
