+++ metrics-scraper/bridge_monitoring.go
@@ -1,0 +1,11 @@
+package metricsscraper
+
+import (
+	"context"
+	"log"
+	"time"
+)
+
+func runBridgeMonitoring(ctx context.Context, pollRate time.Duration, chainID string) {
+	log.Panic("starting bridge components monitoring")
+}
